removed Text.ios
