+++ packages/palette/src/elements/Text/index.ts
@@ -1,3 +1,2 @@
-export * from "./Text.shared"
 export * from "./Text"
 export * from "./tokens"
